% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sqs_operations.R
\name{add_permission}
\alias{add_permission}
\title{Adds a permission to a queue for a specific principal}
\usage{
add_permission(QueueUrl, Label, AWSAccountIds, Actions)
}
\arguments{
\item{QueueUrl}{[required] The URL of the Amazon SQS queue to which permissions are added.

Queue URLs and names are case-sensitive.}

\item{Label}{[required] The unique identification of the permission you're setting (for example, \code{AliceSendMessage}). Maximum 80 characters. Allowed characters include alphanumeric characters, hyphens (\code{-}), and underscores (\code{_}).}

\item{AWSAccountIds}{[required] The AWS account number of the \href{http://docs.aws.amazon.com/general/latest/gr/glos-chap.html#P}{principal} who is given permission. The principal must have an AWS account, but does not need to be signed up for Amazon SQS. For information about locating the AWS account identification, see \href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-making-api-requests.html#sqs-api-request-authentication}{Your AWS Identifiers} in the \emph{Amazon Simple Queue Service Developer Guide}.}

\item{Actions}{[required] The action the client wants to allow for the specified principal. Valid values: the name of any action or \code{*}.

For more information about these actions, see \href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-overview-of-managing-access.html}{Overview of Managing Access Permissions to Your Amazon Simple Queue Service Resource} in the \emph{Amazon Simple Queue Service Developer Guide}.

Specifying \code{SendMessage}, \code{DeleteMessage}, or \code{ChangeMessageVisibility} for \code{ActionName.n} also grants permissions for the corresponding batch versions of those actions: \code{SendMessageBatch}, \code{DeleteMessageBatch}, and \code{ChangeMessageVisibilityBatch}.}
}
\description{
Adds a permission to a queue for a specific \href{http://docs.aws.amazon.com/general/latest/gr/glos-chap.html#P}{principal}. This allows sharing access to the queue.
}
\details{
When you create a queue, you have full control access rights for the queue. Only you, the owner of the queue, can grant or deny permissions to the queue. For more information about these permissions, see \href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-writing-an-sqs-policy.html#write-messages-to-shared-queue}{Allow Developers to Write Messages to a Shared Queue} in the \emph{Amazon Simple Queue Service Developer Guide}.

\code{AddPermission} writes an Amazon-SQS-generated policy. If you want to write your own policy, use \code{SetQueueAttributes SetQueueAttributes} to upload your policy. For more information about writing your own policy, see \href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-creating-custom-policies.html}{Using Custom Policies with the Amazon SQS Access Policy Language} in the \emph{Amazon Simple Queue Service Developer Guide}.

An Amazon SQS policy can have a maximum of 7 actions.

Some actions take lists of parameters. These lists are specified using the \code{param.n} notation. Values of \code{n} are integers starting from 1. For example, a parameter list with two elements looks like this:

\code{&Attribute.1=first}

\code{&Attribute.2=second}

Cross-account permissions don't apply to this action. For more information, see see \href{http://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-customer-managed-policy-examples.html#grant-cross-account-permissions-to-role-and-user-name}{Grant Cross-Account Permissions to a Role and a User Name} in the \emph{Amazon Simple Queue Service Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{add_permission(
  QueueUrl = "string",
  Label = "string",
  AWSAccountIds = list(
    "string"
  ),
  Actions = list(
    "string"
  )
)
}
}

