% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sqs_operations.R
\name{change_message_visibility_batch}
\alias{change_message_visibility_batch}
\title{Changes the visibility timeout of multiple messages}
\usage{
change_message_visibility_batch(QueueUrl, Entries)
}
\arguments{
\item{QueueUrl}{The URL of the Amazon SQS queue whose messages' visibility is changed.

Queue URLs and names are case-sensitive.}

\item{Entries}{A list of receipt handles of the messages for which the visibility timeout must be changed.}
}
\description{
Changes the visibility timeout of multiple messages. This is a batch version of \code{ ChangeMessageVisibility.ChangeMessageVisibility.} The result of the action on each message is reported individually in the response. You can send up to 10 \code{ChangeMessageVisibility ChangeMessageVisibility} requests with each \code{ChangeMessageVisibilityBatch} action.
}
\details{
Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of \code{200}.

Some actions take lists of parameters. These lists are specified using the \code{param.n} notation. Values of \code{n} are integers starting from 1. For example, a parameter list with two elements looks like this:

\code{&Attribute.1=first}

\code{&Attribute.2=second}
}
\examples{

}
