% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sqs_operations.R
\name{delete_message_batch}
\alias{delete_message_batch}
\title{Deletes up to ten messages from the specified queue}
\usage{
delete_message_batch(QueueUrl, Entries)
}
\arguments{
\item{QueueUrl}{The URL of the Amazon SQS queue from which messages are deleted.

Queue URLs and names are case-sensitive.}

\item{Entries}{A list of receipt handles for the messages to be deleted.}
}
\description{
Deletes up to ten messages from the specified queue. This is a batch version of \code{ DeleteMessage.DeleteMessage.} The result of the action on each message is reported individually in the response.
}
\details{
Because the batch request can result in a combination of successful and unsuccessful actions, you should check for batch errors even when the call returns an HTTP status code of \code{200}.

Some actions take lists of parameters. These lists are specified using the \code{param.n} notation. Values of \code{n} are integers starting from 1. For example, a parameter list with two elements looks like this:

\code{&Attribute.1=first}

\code{&Attribute.2=second}
}
\examples{

}
