% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sqs_package.R
\docType{package}
\name{paws.sqs-package}
\alias{paws.sqs}
\alias{paws.sqs-package}
\title{paws.sqs: Amazon Simple Queue Service}
\description{
Welcome to the *Amazon Simple Queue Service API
Reference*.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
